Route Refactor
Refactor routes to move code into their own middleware
