+++ routes/reviews.js
@@ -1,21 +1,10 @@
 const express = require('express');
 const router = express.Router({ mergeParams: true });
-const catchAsync = require('../utils/catchAsync');
-const ExpressError = require('../utils/ExpressError');
+const { validateReview } = require('../middleware');
 const Campground = require('../models/campground');
 const Review = require('../models/review');
-
-const { reviewSchema } = require('../schemas.js');
-
-const validateReview = (req, res, next) => {
-    const { error } = reviewSchema.validate(req.body);
-    if (error) {
-        const msg = error.details.map(el => el.message).join(',');
-        throw new ExpressError(msg, 400);
-    } else {
-        next();
-    }
-}
+const ExpressError = require('../utils/ExpressError');
+const catchAsync = require('../utils/catchAsync');
 
 router.post('/', validateReview, catchAsync(async (req, res) => {
     const campground = await Campground.findById(req.params.id);
